requirements.txt
fastapi==0.115.2
pydantic==2.9.2
langchain==0.3.4
langchain-openai==0.2.3
langchain-core==0.3.13
python-dotenv==1.0.1
typing_extensions==4.12.2
uvicorn==0.32.0
gensim==4.3.3
aspose-pdf==24.10.0
PyMuPDF==1.24.13